Split the shell bundle to fix vite's >500kB chunk-size warning
Route-gated surfaces that never render on the explorer front door
(Preferences, Templates, Mounts, the app-builder hub, Claude Config,
Claude Artifacts, bookmark-open) now load via React.lazy + Suspense
instead of being statically imported into shell/App.tsx. Split
apps/claude_config's useClaudeConfigAvailable probe out of its index.ts
barrel into available.ts so ShellSidebar's eager availability check
doesn't drag the (now-lazy) settings panel back into the main chunk.
Also split react/react-dom/driver.js into their own vendor chunk via
vite's manualChunks.

The largest emitted chunk drops from 585 kB to 273 kB, and the build
warning is gone.
